modifying/adding client side error handling conditions in all ajax calls in PA

+++ angularjs/ui2/src/app/pa/dashboard/abstraction/scientific/pasEligibilityCtrl.js
@@ -65,7 +65,7 @@
             outerTrial.trial.lock_version = PATrialService.getCurrentTrialFromCache().lock_version;
             PATrialService.upsertTrial(outerTrial).then(function(res) {
 
-                if (res.server_response.status === 200) {
+                if (res.server_response.status >= 200 && status <= 210) {
                     vm.trialDetailObj = res;
                     vm.trialDetailObj.lock_version = res.lock_version;
                     vm.deleteAllOCCheckbox = false;
